+++ templates/Users/profile.php
@@ -32,7 +32,7 @@
             <p><?= h($user->username) ?></p>
             <h6 class="subheader"><?= __d('cake_d_c/users', 'Email') ?></h6>
             <p><?= h($user->email) ?></p>
-            <?= $this->User->socialConnectLinkList($user->social_accounts) ?>
+            <?= $this->User->socialConnectLinkList($user->social_accounts ?? []) ?>
             <?php
             if (!empty($user->social_accounts)):
                 ?>