+++ tests/TestCase/Controller/Traits/Webauthn2FaTraitTest.php
@@ -108,11 +108,10 @@ class Webauthn2FaTraitTest extends BaseTrait
         $this->Trait
             ->expects($this->exactly(2))
             ->method('set')
-            ->willReturnCallback(fn ($name, $value) =>
-                match ([$name, $value]) {
-                    ['isRegister', true] => null,
-                    ['username', 'user-2'] => null
-                });
+            ->willReturnCallback(fn ($name, $value) => match ([$name, $value]) {
+                ['isRegister', true] => null,
+                ['username', 'user-2'] => null
+            });
         $this->Trait->webauthn2fa();
         $this->assertSame(
             $user,
@@ -146,11 +145,10 @@ class Webauthn2FaTraitTest extends BaseTrait
         $this->Trait
             ->expects($this->exactly(2))
             ->method('set')
-            ->willReturnCallback(fn ($name, $value) =>
-                match ([$name, $value]) {
-                    ['isRegister', false] => null,
-                    ['username', 'user-1'] => null
-                });
+            ->willReturnCallback(fn ($name, $value) => match ([$name, $value]) {
+                ['isRegister', false] => null,
+                ['username', 'user-1'] => null
+            });
         $this->Trait->webauthn2fa();
         $this->assertSame(
             $user,