+++ webroot/js/reCaptchaV3.js
@@ -1,0 +1,3 @@
+function onSubmit(token) {
+    document.forms[0].submit()
+}